fixing renames and position to match scaffolding
