+++ routes/entries.js
@@ -1,17 +1,12 @@
 const express = require('express')
-
-const { createEntry } = require('../controllers/news')
-const { validateCreate } = require('../validators/entries')
-
-const router = express.Router()
-
-router.post('/', validateCreate, createEntry)
-
-const { get } = require('../controllers/entries')
+const { get } = require('../controllers/index')
+const { post } = require('../controllers/entries')
+const { validateCreate } = require('../schemas/entries')
 
 const router = express.Router()
+
+router.post('/', validateCreate, post)
 
 router.get('/', get)
-
 
 module.exports = router
